+++ CHANGELOG.md
@@ -43,7 +43,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix return to previous page on screen size change - #710 by @orzechdev
 - Add variants reordering possibility - #716 by @orzechdev
 - Fix avatar change button - #719 by @orzechdev
-- Default product variant - #721 by @tomaszszymanski129
+- Allow product variant to be set as default - #721 by @tomaszszymanski129
 
 ## 2.10.1